// src/utils/dto/sidebar-product.dto.ts
export interface SideBarCarsProductsDto{
    vehicleName?: string
    vehicleType?: string
    sort?: string
    brandName?: string[]
    transmission?: string[]
    seatsNum?: number[]
    minPrice?: number
    maxPrice?: number
    minPower?: number
    maxPower?: number
    location?: string[]
}